Renames in dataset creation code.
This PR just renames things, will add tutorial and move Beaker stuff
around in a separate PR.
