+++ olmoearth_pretrain/dataset_creation/make_meta_summary.py
@@ -12,7 +12,7 @@ from .util import get_modality_dir, get_modality_temp_meta_dir
 
 
 def make_meta_summary(
-    helios_path: UPath, modality: ModalitySpec, time_span: TimeSpan
+    olmoearth_path: UPath, modality: ModalitySpec, time_span: TimeSpan
 ) -> None:
     """Create the concatenated metadata file for the specified modality.
 
@@ -21,7 +21,7 @@ def make_meta_summary(
     together into one big CSV.
 
     Args:
-        helios_path: OlmoEarth Pretrain dataset path.
+        olmoearth_path: OlmoEarth Pretrain dataset path.
         modality: modality to write summary for.
         time_span: time span to write summary for.
     """
@@ -29,8 +29,8 @@ def make_meta_summary(
     # read.
     column_names: list[str] | None = None
     csv_rows = []
-    modality_dir = get_modality_dir(helios_path, modality, time_span)
-    meta_dir = get_modality_temp_meta_dir(helios_path, modality, time_span)
+    modality_dir = get_modality_dir(olmoearth_path, modality, time_span)
+    meta_dir = get_modality_temp_meta_dir(olmoearth_path, modality, time_span)
     meta_fnames = list(meta_dir.iterdir())
     for fname in tqdm.tqdm(meta_fnames):
         with fname.open() as f:
@@ -45,7 +45,7 @@ def make_meta_summary(
     if column_names is None:
         raise ValueError(f"did not find any files in {meta_dir}")
 
-    with (helios_path / f"{modality_dir.name}.csv").open("w") as f:
+    with (olmoearth_path / f"{modality_dir.name}.csv").open("w") as f:
         writer = csv.DictWriter(f, fieldnames=column_names)
         writer.writeheader()
         writer.writerows(csv_rows)
@@ -56,7 +56,7 @@ if __name__ == "__main__":
         description="Create concatenated metadata file",
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="OlmoEarth Pretrain dataset path",
         required=True,
@@ -76,4 +76,4 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     modality = Modality.get(args.modality)
-    make_meta_summary(UPath(args.helios_path), modality, TimeSpan(args.time_span))
+    make_meta_summary(UPath(args.olmoearth_path), modality, TimeSpan(args.time_span))
