+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/era5.py
@@ -27,12 +27,12 @@ LAYER_NAME = "era5"
 logger = logging.getLogger(__name__)
 
 
-def convert_era5(window_path: UPath, helios_path: UPath) -> None:
+def convert_era5(window_path: UPath, olmoearth_path: UPath) -> None:
     """Add ERA5 data for this window to the OlmoEarth Pretrain dataset.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     modality = Modality.ERA5
     assert len(modality.band_sets) == 1
@@ -109,7 +109,7 @@ def convert_era5(window_path: UPath, helios_path: UPath) -> None:
     # Save the one-year image and metadata.
     year_stacked_image = np.concatenate(year_images, axis=0)
     year_dst_fname = get_modality_fname(
-        helios_path,
+        olmoearth_path,
         Modality.ERA5,
         TimeSpan.YEAR,
         window_metadata,
@@ -124,7 +124,7 @@ def convert_era5(window_path: UPath, helios_path: UPath) -> None:
         fname=year_dst_fname.name,
     )
     year_metadata_fname = get_modality_temp_meta_fname(
-        helios_path, Modality.ERA5, TimeSpan.YEAR, window.name
+        olmoearth_path, Modality.ERA5, TimeSpan.YEAR, window.name
     )
     year_metadata_fname.parent.mkdir(parents=True, exist_ok=True)
     with year_metadata_fname.open("w") as f:
@@ -145,7 +145,7 @@ def convert_era5(window_path: UPath, helios_path: UPath) -> None:
 
     # Save the two-week image and metadata.
     two_week_dst_fname = get_modality_fname(
-        helios_path,
+        olmoearth_path,
         Modality.ERA5,
         TimeSpan.TWO_WEEK,
         window_metadata,
@@ -160,7 +160,7 @@ def convert_era5(window_path: UPath, helios_path: UPath) -> None:
         fname=two_week_dst_fname.name,
     )
     two_week_metadata_fname = get_modality_temp_meta_fname(
-        helios_path, Modality.ERA5, TimeSpan.TWO_WEEK, window.name
+        olmoearth_path, Modality.ERA5, TimeSpan.TWO_WEEK, window.name
     )
     two_week_metadata_fname.parent.mkdir(parents=True, exist_ok=True)
     with two_week_metadata_fname.open("w") as f:
@@ -192,7 +192,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -206,7 +206,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
     metadata_fnames = ds_path.glob("windows/res_160/*/metadata.json")
     jobs = []
@@ -214,7 +214,7 @@ if __name__ == "__main__":
         jobs.append(
             dict(
                 window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                olmoearth_path=olmoearth_path,
             )
         )
