+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/naip_10.py
@@ -22,14 +22,14 @@ from .multitemporal_raster import get_adjusted_projection_and_bounds
 LAYER_NAME = "naip_10"
 
 
-def convert_naip(window_path: UPath, helios_path: UPath) -> None:
+def convert_naip(window_path: UPath, olmoearth_path: UPath) -> None:
     """Add NAIP data for this window to the OlmoEarth Pretrain dataset.
 
     This is for NAIP data at 4096 x 4096 under 10 m/pixel tiling.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     window = Window.load(window_path)
     window_metadata = get_window_metadata(window)
@@ -68,7 +68,7 @@ def convert_naip(window_path: UPath, helios_path: UPath) -> None:
         raster_dir, adjusted_projection, adjusted_bounds
     )
     dst_fname = get_modality_fname(
-        helios_path,
+        olmoearth_path,
         Modality.NAIP_10,
         TimeSpan.STATIC,
         window_metadata,
@@ -83,7 +83,7 @@ def convert_naip(window_path: UPath, helios_path: UPath) -> None:
         fname=dst_fname.name,
     )
     metadata_fname = get_modality_temp_meta_fname(
-        helios_path, Modality.NAIP_10, TimeSpan.STATIC, window.name
+        olmoearth_path, Modality.NAIP_10, TimeSpan.STATIC, window.name
     )
     metadata_fname.parent.mkdir(parents=True, exist_ok=True)
     with metadata_fname.open("w") as f:
@@ -115,7 +115,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -129,7 +129,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
     metadata_fnames = ds_path.glob("windows/res_10/*/metadata.json")
     jobs = []
@@ -137,7 +137,7 @@ if __name__ == "__main__":
         jobs.append(
             dict(
                 window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                olmoearth_path=olmoearth_path,
             )
         )
