+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/openstreetmap.py
@@ -36,12 +36,12 @@ LAYER_NAME = "openstreetmap"
 RESOLUTION = 10
 
 
-def convert_openstreetmap(window_path: UPath, helios_path: UPath) -> None:
+def convert_openstreetmap(window_path: UPath, olmoearth_path: UPath) -> None:
     """Add OpenStreetMap data for this window to the OlmoEarth Pretrain dataset.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     window = Window.load(window_path)
     layer_datas = window.load_layer_datas()
@@ -67,7 +67,7 @@ def convert_openstreetmap(window_path: UPath, helios_path: UPath) -> None:
 
     # Upload the data.
     dst_fname = get_modality_fname(
-        helios_path,
+        olmoearth_path,
         Modality.OPENSTREETMAP,
         TimeSpan.STATIC,
         window_metadata,
@@ -82,7 +82,7 @@ def convert_openstreetmap(window_path: UPath, helios_path: UPath) -> None:
 
     # Create the metadata file for this data.
     metadata_fname = get_modality_temp_meta_fname(
-        helios_path, Modality.OPENSTREETMAP, TimeSpan.STATIC, window.name
+        olmoearth_path, Modality.OPENSTREETMAP, TimeSpan.STATIC, window.name
     )
     metadata_fname.parent.mkdir(parents=True, exist_ok=True)
     with metadata_fname.open("w") as f:
@@ -114,7 +114,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -128,7 +128,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
     metadata_fnames = ds_path.glob("windows/*/*/metadata.json")
     jobs = []
@@ -136,7 +136,7 @@ if __name__ == "__main__":
         jobs.append(
             dict(
                 window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                olmoearth_path=olmoearth_path,
             )
         )
