+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/sentinel1.py
@@ -18,17 +18,17 @@ LAYER_FREQ = "sentinel1_freq"
 LAYER_MONTHLY = "sentinel1"
 
 
-def convert_sentinel1(window_path: UPath, helios_path: UPath) -> None:
+def convert_sentinel1(window_path: UPath, olmoearth_path: UPath) -> None:
     """Add Landsat data for this window to the OlmoEarth Pretrain dataset.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     try:
         convert_freq(
             window_path,
-            helios_path,
+            olmoearth_path,
             LAYER_FREQ,
             Modality.SENTINEL1,
             missing_okay=True,
@@ -40,7 +40,7 @@ def convert_sentinel1(window_path: UPath, helios_path: UPath) -> None:
         )
 
     try:
-        convert_monthly(window_path, helios_path, LAYER_MONTHLY, Modality.SENTINEL1)
+        convert_monthly(window_path, olmoearth_path, LAYER_MONTHLY, Modality.SENTINEL1)
     except Exception as e:
         print(
             f"warning: got error {e} while converting monthly data for window {window_path}"
@@ -60,7 +60,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -74,7 +74,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
     metadata_fnames = ds_path.glob("windows/res_10/*/metadata.json")
     jobs = []
@@ -82,7 +82,7 @@ if __name__ == "__main__":
         jobs.append(
             dict(
                 window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                olmoearth_path=olmoearth_path,
             )
         )
