+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/sentinel2.py
@@ -1,4 +1,4 @@
-"""Post-process ingested Landsat data into the OlmoEarth Pretrain dataset."""
+"""Post-process ingested Sentinel-2 data into the OlmoEarth Pretrain dataset."""
 
 import argparse
 import multiprocessing
@@ -12,27 +12,28 @@ from olmoearth_pretrain.data.constants import Modality
 from .multitemporal_raster import convert_freq, convert_monthly
 
 # rslearn layer for frequent data.
-LAYER_FREQ = "landsat_freq"
+LAYER_FREQ = "sentinel2_freq"
 
 # rslearn layer prefix for monthly data.
-LAYER_MONTHLY = "landsat"
+LAYER_MONTHLY = "sentinel2"
 
 
-def convert_landsat(window_path: UPath, helios_path: UPath) -> None:
-    """Add Landsat data for this window to the OlmoEarth Pretrain dataset.
+def convert_sentinel2(window_path: UPath, olmoearth_path: UPath) -> None:
+    """Add Sentinel-2 data for this window to the OlmoEarth Pretrain dataset.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     convert_freq(
         window_path,
-        helios_path,
+        olmoearth_path,
         LAYER_FREQ,
-        Modality.LANDSAT,
+        Modality.SENTINEL2,
         missing_okay=True,
+        unprepared_okay=True,
     )
-    convert_monthly(window_path, helios_path, LAYER_MONTHLY, Modality.LANDSAT)
+    convert_monthly(window_path, olmoearth_path, LAYER_MONTHLY, Modality.SENTINEL2)
 
 
 if __name__ == "__main__":
@@ -48,7 +49,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -62,20 +63,20 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
-    metadata_fnames = ds_path.glob("windows/res_10/*/metadata.json")
     jobs = []
-    for metadata_fname in metadata_fnames:
+    group_dir = ds_path / "windows" / "res_10"
+    for window_dir in group_dir.iterdir():
         jobs.append(
             dict(
-                window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                window_path=window_dir,
+                olmoearth_path=olmoearth_path,
             )
         )
 
     p = multiprocessing.Pool(args.workers)
-    outputs = star_imap_unordered(p, convert_landsat, jobs)
+    outputs = star_imap_unordered(p, convert_sentinel2, jobs)
     for _ in tqdm.tqdm(outputs, total=len(jobs)):
         pass
     p.close()
