+++ olmoearth_pretrain/dataset_creation/rslearn_to_olmoearth/srtm.py
@@ -1,4 +1,4 @@
-"""Post-process ingested WorldCover data into the OlmoEarth Pretrain dataset."""
+"""Post-process ingested SRTM elevation data into the OlmoEarth Pretrain dataset."""
 
 import argparse
 import csv
@@ -16,19 +16,19 @@ from olmoearth_pretrain.dataset.utils import get_modality_fname
 from ..constants import GEOTIFF_RASTER_FORMAT, METADATA_COLUMNS
 from ..util import get_modality_temp_meta_fname, get_window_metadata
 
-START_TIME = datetime(2021, 1, 1, tzinfo=UTC)
-END_TIME = datetime(2022, 1, 1, tzinfo=UTC)
+START_TIME = datetime(2000, 1, 1, tzinfo=UTC)
+END_TIME = datetime(2001, 1, 1, tzinfo=UTC)
 
 # Layer name in the input rslearn dataset.
-LAYER_NAME = "worldcover"
+LAYER_NAME = "srtm"
 
 
-def convert_worldcover(window_path: UPath, helios_path: UPath) -> None:
-    """Add WorldCover data for this window to the OlmoEarth Pretrain dataset.
+def convert_srtm(window_path: UPath, olmoearth_path: UPath) -> None:
+    """Add SRTM elevation data for this window to the OlmoEarth Pretrain dataset.
 
     Args:
         window_path: the rslearn window directory to read data from.
-        helios_path: OlmoEarth Pretrain dataset path to write to.
+        olmoearth_path: OlmoEarth Pretrain dataset path to write to.
     """
     window = Window.load(window_path)
     window_metadata = get_window_metadata(window)
@@ -36,15 +36,15 @@ def convert_worldcover(window_path: UPath, helios_path: UPath) -> None:
     if not window.is_layer_completed(LAYER_NAME):
         return
 
-    assert len(Modality.WORLDCOVER.band_sets) == 1
-    band_set = Modality.WORLDCOVER.band_sets[0]
+    assert len(Modality.SRTM.band_sets) == 1
+    band_set = Modality.SRTM.band_sets[0]
     raster_dir = window.get_raster_dir(LAYER_NAME, band_set.bands)
     image = GEOTIFF_RASTER_FORMAT.decode_raster(
         raster_dir, window.projection, window.bounds
     )
     dst_fname = get_modality_fname(
-        helios_path,
-        Modality.WORLDCOVER,
+        olmoearth_path,
+        Modality.SRTM,
         TimeSpan.STATIC,
         window_metadata,
         band_set.get_resolution(),
@@ -58,7 +58,7 @@ def convert_worldcover(window_path: UPath, helios_path: UPath) -> None:
         fname=dst_fname.name,
     )
     metadata_fname = get_modality_temp_meta_fname(
-        helios_path, Modality.WORLDCOVER, TimeSpan.STATIC, window.name
+        olmoearth_path, Modality.SRTM, TimeSpan.STATIC, window.name
     )
     metadata_fname.parent.mkdir(parents=True, exist_ok=True)
     with metadata_fname.open("w") as f:
@@ -90,7 +90,7 @@ if __name__ == "__main__":
         required=True,
     )
     parser.add_argument(
-        "--helios_path",
+        "--olmoearth_path",
         type=str,
         help="Destination OlmoEarth Pretrain dataset path",
         required=True,
@@ -104,20 +104,19 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     ds_path = UPath(args.ds_path)
-    helios_path = UPath(args.helios_path)
+    olmoearth_path = UPath(args.olmoearth_path)
 
-    metadata_fnames = ds_path.glob("windows/*/*/metadata.json")
     jobs = []
-    for metadata_fname in metadata_fnames:
+    for window_dir in (ds_path / "windows" / "res_10").iterdir():
         jobs.append(
             dict(
-                window_path=metadata_fname.parent,
-                helios_path=helios_path,
+                window_path=window_dir,
+                olmoearth_path=olmoearth_path,
             )
         )
 
     p = multiprocessing.Pool(args.workers)
-    outputs = star_imap_unordered(p, convert_worldcover, jobs)
+    outputs = star_imap_unordered(p, convert_srtm, jobs)
     for _ in tqdm.tqdm(outputs, total=len(jobs)):
         pass
     p.close()
